% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{as.data.frame.GSON}
\alias{as.data.frame.GSON}
\title{Coerce GSON to a data frame}
\usage{
\method{as.data.frame}{GSON}(x, row.names = NULL, optional = FALSE, ...)
}
\arguments{
\item{x}{A \code{GSON} object.}

\item{row.names}{Unused.}

\item{optional}{Unused.}

\item{...}{Unused.}
}
\value{
A data frame with gene set-gene memberships and optional names.
}
\description{
Coerce GSON to a data frame
}
