% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{[[.GSON}
\alias{[[.GSON}
\title{Extract genes from a GSON object}
\usage{
\method{[[}{GSON}(x, i, ...)
}
\arguments{
\item{x}{A \code{GSON} object.}

\item{i}{A gene set ID or numeric index.}

\item{...}{Unused.}
}
\value{
A character vector of genes.
}
\description{
Extract genes from a GSON object
}
