% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/go_mapper.R
\name{go_mapper}
\alias{go_mapper}
\title{Build a gson object that annotate Gene Ontology}
\usage{
go_mapper(data, ont = c("ALL", "BP", "CC", "MF"), species = NULL, ...)
}
\arguments{
\item{data}{a two-column data.frame of original GO annotation. The columns are "gene_id" and "go_id".}

\item{ont}{type of GO annotation, which is "ALL", "BP", "MF", or "CC". default: "ALL".}

\item{species}{name of species. Default: NULL.}

\item{...}{pass to \code{gson::gson()} constructor.}
}
\value{
a \code{gson} instance
}
\description{
Build a gson object that annotate Gene Ontology
}
\examples{
 data = data.frame(gene_id = "gene1", 
                   go_id = c("GO:0035492", "GO:0009764", "GO:0031063", "GO:0033714", "GO:0036349"))
 go_mapper(data, species = "E. coli")
}
