% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-AllClasses.R
\docType{class}
\name{GSON-class}
\alias{GSON-class}
\alias{show,GSON-method}
\title{Class "GSON"
This class represents gene set information.}
\description{
Class "GSON"
This class represents gene set information.
}
\section{Slots}{

\describe{
\item{\code{gsid2gene}}{data.frame with two columns of 'gsid' and 'gene'}

\item{\code{gsid2name}}{data.frame with two columns of 'gsid' and 'name'}

\item{\code{gene2name}}{data.frame with two columns of 'gene' and 'name'}

\item{\code{schema_version}}{version of the GSON file schema}

\item{\code{species}}{species of the annotation}

\item{\code{gsname}}{gene set name, e.g., GO, KEGG}

\item{\code{version}}{version of the gene set}

\item{\code{accessed_date}}{time to obtain the gene set data}

\item{\code{keytype}}{keytype of genes}

\item{\code{urlpattern}}{URL pattern to browse gene set online}

\item{\code{info}}{extra information}
}}

\author{
Guangchuang Yu \url{https://yulab-smu.top}
}
\keyword{classes}
