% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsonList.R
\name{gsonList}
\alias{gsonList}
\title{construct a 'GSONList' object}
\usage{
gsonList(...)
}
\arguments{
\item{...}{input GSON objects}
}
\value{
A 'GSONList' instance
}
\description{
construct a 'GSONList' object
}
