% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IO.R
\name{read.gson}
\alias{read.gson}
\alias{write.gson}
\title{read and write gson file}
\usage{
read.gson(file)

write.gson(x, file = "")
}
\arguments{
\item{file}{A gson file}

\item{x}{A \code{GSON} instance}
}
\value{
A \code{GSON} instance
}
\description{
read and write gson file
}
\examples{
wpfile <- system.file('extdata', "wikipathways-20220310-gmt-Homo_sapiens.gmt", package='gson')
x <- read.gmt.wp(wpfile, output = "GSON")
f = tempfile(fileext = '.gson')
write.gson(x, f)
read.gson(f)
}
