% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kegg_mapper.R
\name{kegg_mapper}
\alias{kegg_mapper}
\title{Build KEGG annotation for novel species using KEGG Mapper
KEGG Mapper service can annotate protein sequences for novel species with KO database,
and KO annotation need to be converted into Pathway or Module annotation, which
can then be used in \code{clusterProfiler}}
\usage{
kegg_mapper(
  file,
  format = c("BLAST", "Ghost", "Kofam"),
  type = c("pathway", "module"),
  species = NULL,
  ...
)
}
\arguments{
\item{file}{the name of the file which comes from the KEGG Mapper service}

\item{format}{string indicate format of KEGG Mapper result}

\item{type}{string indicate annotation database}

\item{species}{your species, NULL if ignored}

\item{...}{pass to gson::gson()}
}
\value{
a gson instance
}
\description{
Build KEGG annotation for novel species using KEGG Mapper
KEGG Mapper service can annotate protein sequences for novel species with KO database,
and KO annotation need to be converted into Pathway or Module annotation, which
can then be used in \code{clusterProfiler}
}
\examples{
\dontrun{
 file = system.file('extdata', "kegg_mapper_blast.txt", package='gson')
 kegg_mapper(file, format = "BLAST", type = "pathway")
}
}
