% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_gson.R
\name{merge_gson}
\alias{merge_gson}
\title{Merge a list of \code{gson} object}
\usage{
merge_gson(..., glist)
}
\arguments{
\item{...}{additional \code{gson} object}

\item{glist}{a list of \code{gson} objects}
}
\value{
a \code{gson} instance
}
\description{
Merge a list of \code{gson} object
}
\note{
Do not merge \code{gson} objects of different annotation types
}
\examples{
\dontrun{
  merge_gson(gson1, gson2, gson3)
}
}
