% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate_gene_id.R
\name{mutate_gene_id}
\alias{mutate_gene_id}
\title{Transform the gene id in \code{gson} object}
\usage{
mutate_gene_id(x, .f, ...)
}
\arguments{
\item{x}{a json object}

\item{.f}{a function fun, a quosure style lambda ~ fun(.) or a list of either form.}

\item{...}{additional arguments for the function call in .f.}
}
\description{
Transform the gene id in \code{gson} object
}
