% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\docType{methods}
\name{show}
\alias{show}
\title{show method}
\usage{
show(object)
}
\arguments{
\item{object}{A \code{GSON} object}
}
\value{
message
}
\description{
show method for \code{GSON} instance
}
\author{
Guangchuang Yu \url{https://yulab-smu.top}
}
