% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gson-package.R
\docType{package}
\name{gson-package}
\alias{gson-package}
\title{gson: Base Class and Methods for 'gson' Format}
\description{
Proposes a new file format ('gson') for storing gene set and related information, and provides read, write and other utilities to process this file format.
}
\author{
\strong{Maintainer}: Guangchuang Yu \email{guangchuangyu@gmail.com} (\href{https://orcid.org/0000-0002-6485-8781}{ORCID}) [copyright holder]

}
\keyword{internal}
