% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GMT.R
\name{read.gmt}
\alias{read.gmt}
\alias{read.gmt.wp}
\title{read.gmt}
\usage{
read.gmt(gmtfile)

read.gmt.wp(gmtfile, output = "data.frame")
}
\arguments{
\item{gmtfile}{gmt file}

\item{output}{one of 'data.frame' or 'GSON'}
}
\value{
data.frame
}
\description{
parse gmt file to a data.frame
}
\author{
Guangchuang Yu
}
